feat: replace React graph components with imperative GraphView
Replace the React component tree (Graph → GraphCanvas + Labels) with
an imperative GraphView class that directly manages the WebGL canvas,
label DOM overlay, camera, and interaction handler.

Key changes:
- GraphViewComputed.ts: pure computation functions extracted from
  useMemo hooks (node colors, sizes, edge colors, arrows)
- GraphViewLabels.ts: LabelManager class extracted from Labels.tsx
  with the same DOM management logic
- GraphView.ts: orchestrator class with dirty-flag-gated recomputation,
  on-demand render loop, and throttled label updates
- Graph.tsx: thin React wrapper that creates DOM elements and forwards
  prop changes via setters

This eliminates React re-renders from the pan/zoom hot path entirely.
The render loop, label layout, and all state management are now fully
imperative with no React scheduling overhead.
